Changed PostFactory to class
