+++ src/middleware/logToDB.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from 'express'
-import { LogController } from '../controllers/logToDB'
+import { LogController } from '../controllers/logToDB.controller'
 
 export const logToDB = (req: Request, res: Response, next: NextFunction) => {
     const logController = new LogController(req, res, next)
